fix: complete lib-client u128 migration — zero u64 amounts remaining
Widened ALL amount parameters across lib-client:
- lib.rs JNI wrappers: amount, stable_amount, token_amount, sov_amount,
  min_tokens_out, min_stable_out, amount_in, min_amount_out, lp_amount,
  initial_supply — all u128
- bonding_curve_tx.rs: amount_in, min_amount_out, lp_amount — u128
- Removed all 'as u64' casts on amount values (was truncating 18-decimal)

Remaining u64 fields are non-amounts: nonces, fees, timestamps, byte sizes,
decimals, basis points, curve parameters.

+++ lib-client/src/lib.rs
@@ -1295,7 +1295,7 @@ pub extern "system" fn Java_com_sovereignnetworkmobile_Identity_nativeBuildSovWa
         handle as *const IdentityHandle,
         from_vec.as_ptr(),
         to_vec.as_ptr(),
-        amount as u64,
+        amount as u128,
         chain_id as u8,
         nonce as u64,
     );
@@ -1864,7 +1864,7 @@ pub extern "C" fn zhtp_client_build_bonding_curve_buy(
     handle: *const IdentityHandle,
     token_id: *const u8,
     stable_amount: u128,
-    min_tokens_out: u64,
+    min_tokens_out: u128,
     chain_id: u8,
     nonce: u64,
 ) -> *mut std::ffi::c_char {
@@ -1909,7 +1909,7 @@ pub extern "C" fn zhtp_client_build_bonding_curve_sell(
     handle: *const IdentityHandle,
     token_id: *const u8,
     token_amount: u128,
-    min_stable_out: u64,
+    min_stable_out: u128,
     chain_id: u8,
     nonce: u64,
 ) -> *mut std::ffi::c_char {
@@ -1956,8 +1956,8 @@ pub extern "C" fn zhtp_client_build_swap(
     handle: *const IdentityHandle,
     token_id: *const u8,
     pool_id: *const u8,
-    amount_in: u64,
-    min_amount_out: u64,
+    amount_in: u128,
+    min_amount_out: u128,
     token_to_sov: u8,
     chain_id: u8,
     nonce: u64,
@@ -2059,7 +2059,7 @@ pub extern "C" fn zhtp_client_build_remove_liquidity(
     handle: *const IdentityHandle,
     token_id: *const u8,
     pool_id: *const u8,
-    lp_amount: u64,
+    lp_amount: u128,
     chain_id: u8,
     nonce: u64,
 ) -> *mut std::ffi::c_char {
